// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-linking-selection-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseContextLinkingSelectionAyahDto } from '../models/phrase-context-linking-selection-ayah-dto';
export interface PhraseContextLinkingSelectionResponse {
  activeBuildId: string;
  ayahs: Array<PhraseContextLinkingSelectionAyahDto>;
  selectedAyahCount: number;
}